+++ lib/src/manager/im_manager.dart
@@ -166,6 +166,13 @@ class IMManager {
               messageManager.msgListener
                   .recvMessageExtensionsDeleted(msgID, list);
               break;
+            case 'onRecvMessageExtensionsAdded':
+              var msgID = call.arguments['data']['msgID'];
+              var value = call.arguments['data']['list'];
+              var list = Utils.toList(value, (map) => KeyValue.fromJson(map));
+              messageManager.msgListener
+                  .recvMessageExtensionsAdded(msgID, list);
+              break;
           }
         } else if (call.method == ListenerType.msgSendProgressListener) {
           String type = call.arguments['type'];
@@ -375,6 +382,7 @@ class IMManager {
     String? encryptionKey,
     bool enabledEncryption = false,
     bool enabledCompression = false,
+    bool isExternalExtensions = false,
     String? operationID,
   }) {
     this._connectListener = listener;
@@ -392,6 +400,7 @@ class IMManager {
             "encryption_key": encryptionKey,
             "is_need_encryption": enabledEncryption,
             "is_compression ": enabledCompression,
+            "is_external_extensions": isExternalExtensions,
             "operationID": Utils.checkOperationID(operationID),
           },
         ));